pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "all-weather-strategy"
version = "0.1.0"
description = "全天候期货策略 — 基于风险平价的中国期货市场多资产组合"
requires-python = ">=3.10"
dependencies = [
    "akshare>=1.16.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "pyarrow>=16.0.0",
    "pyyaml>=6.0.1",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
